refactor(backend): extract SAM3 vertical slice
