+++ tests/test_python_package_skeleton.py
@@ -16,6 +16,7 @@ PACKAGE_MODULES = (
     "easyuse_anima.common.serialization",
     "easyuse_anima.image",
     "easyuse_anima.image.geometry",
+    "easyuse_anima.image.sam3",
     "easyuse_anima.infrastructure",
     "easyuse_anima.infrastructure.comfy",
     "easyuse_anima.infrastructure.comfy.capabilities",
@@ -24,9 +25,11 @@ PACKAGE_MODULES = (
     "easyuse_anima.lora",
     "easyuse_anima.naia",
     "easyuse_anima.nodes",
+    "easyuse_anima.nodes.impact_detailer_nodes",
     "easyuse_anima.nodes.prompt_advanced_nodes",
     "easyuse_anima.nodes.prompt_data_nodes",
     "easyuse_anima.nodes.regional_nodes",
+    "easyuse_anima.nodes.sam3_nodes",
     "easyuse_anima.profiles",
     "easyuse_anima.profiles.contract",
     "easyuse_anima.profiles.mutation",
