+++ tests/test_registry_scanner_safety.py
@@ -24,6 +24,7 @@ EXPECTED_PYTHON_PACKAGE_FILES = {
     "easyuse_anima/image/__init__.py",
     "easyuse_anima/image/detailer.py",
     "easyuse_anima/image/geometry.py",
+    "easyuse_anima/image/sam3.py",
     "easyuse_anima/image/scaling.py",
     "easyuse_anima/infrastructure/__init__.py",
     "easyuse_anima/infrastructure/comfy/__init__.py",
@@ -38,10 +39,12 @@ EXPECTED_PYTHON_PACKAGE_FILES = {
     "easyuse_anima/naia/resolution.py",
     "easyuse_anima/nodes/__init__.py",
     "easyuse_anima/nodes/image_nodes.py",
+    "easyuse_anima/nodes/impact_detailer_nodes.py",
     "easyuse_anima/nodes/lora_nodes.py",
     "easyuse_anima/nodes/naia_nodes.py",
     "easyuse_anima/nodes/prompt_data_nodes.py",
     "easyuse_anima/nodes/prompt_nodes.py",
+    "easyuse_anima/nodes/sam3_nodes.py",
     "easyuse_anima/nodes/wildcard_nodes.py",
     "easyuse_anima/profiles/__init__.py",
     "easyuse_anima/profiles/contract.py",
@@ -150,16 +153,19 @@ class RegistryScannerSafetyTests(unittest.TestCase):
             "easyuse_anima/infrastructure/comfy/invocation.py",
             "easyuse_anima/infrastructure/comfy/resources.py",
             "easyuse_anima/image/detailer.py",
+            "easyuse_anima/image/sam3.py",
             "easyuse_anima/image/scaling.py",
             "easyuse_anima/lora/metadata.py",
             "easyuse_anima/lora/preset.py",
             "easyuse_anima/naia/client.py",
             "easyuse_anima/naia/resolution.py",
             "easyuse_anima/nodes/image_nodes.py",
+            "easyuse_anima/nodes/impact_detailer_nodes.py",
             "easyuse_anima/nodes/lora_nodes.py",
             "easyuse_anima/nodes/naia_nodes.py",
             "easyuse_anima/nodes/prompt_data_nodes.py",
             "easyuse_anima/nodes/prompt_nodes.py",
+            "easyuse_anima/nodes/sam3_nodes.py",
             "easyuse_anima/nodes/wildcard_nodes.py",
             "easyuse_anima/prompt/correction.py",
             "easyuse_anima/prompt/artist_mix.py",
